requirements.txt
moviepy==1.0.3
Pillow==10.1.0
requests==2.31.0
numpy==1.26.2
pygame==2.5.2
tqdm==4.66.1
proglog==0.1.10
imageio==2.33.0
imageio-ffmpeg==0.4.9
decorator==4.4.2
